Rk3 (#4)
* cors implemented

* signup validation refactored

* CRUD

* small fix

* signup + autologin

* from HashMap to real Database

* Tests and fixes

* fix

* fix

* fix

* fix

* for heroku

* for heroku 2

* .travis.yml

* .travis.yml fix 1

* 2 more tests

* auto deploy config

* auto deploy config

* auto deploy config

* auto deploy config

* auto deploy config finally (may be)

* auto deploy config finally (may be)

* auto deploy config finally (may be this time)

* without auto deploy. gave up for a while..

* EnableTransactionManagement

* MapMeta added

* checkstyle fix

* Helpers and maps creation added

* commented code deleted

* refactoring

* tests fixed

* pgtypes fixed

* postgres addon 9.6 added

* Domains renamed. Extra functionality added

* Domains renamed. Extra functionality added

* game data to new branch

* fixes

* fixes

* api added

* World Parser added

* validation refactored

* tests fixed

* updated spring-boot version

* some javadoc added

* websocket test

* travis file modified

* travis file modified

* travis

* travis

* travis

* cors temporarily disabled

* cors temporarily disabled

* fixes

* fixes

* web socket testing

* fixes

* fixes

* fixes

* fixes

* fixes

* fix

* fix

* fixes

* fixes

* fixes

* fix

* fix

* fixes

* fix

* fixes

* fixes

* fixes

* fixes

* fixes

* fixes

* logging

* logging

* logging

* logging

* logging

* log

* logging

* logging

* logging

* logging

* fix may be

* yes - it was fix

* Причесал

* logging

* logging fix

* fix

* fix

* new

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* Update README.md

* fix

* fix

* good fix

* good fix

* good fix

* good fix

* good fix

* good fix

* good fix

* good fix

* good fix

* good fix

* nice fix

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* oauth testing

* new oauth testing

* new oauth testing

* new oauth testing

* persisting new vk user

* fix

* fix

* testing

* testing

* testing

* cleaning up. oauth with vk is done

* checkstyle fix

* pom file clean up

* cors extended

* cors extended

* findUserByEmail() method is now covered

* merged master

* rk3

* cors testing

* cors reduced

* cors fixing

* cors fixed

* requested fixes

+++ src/main/resources/db/migration/V3__mapmeta.sql
@@ -1,6 +1,6 @@
 CREATE TABLE board_meta (
   id           BIGSERIAL PRIMARY KEY,
-  board_id     INT,
+  board_id     BIGINT,
   name         TEXT UNIQUE,
   level        INT,
   timer        INT,
